Uploaded V1 of all short SVGs
I decided to go ahead and make all the short SVGs and map them all out! I'm wondering if I added some of the symbols and mapped some of the stuff correctly though, particularly for the obstructed areas and the water. Let me know what you think and if it's good we can go ahead and merge this branch!